requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
jwt
python-multipart